Fix `/problem` web API
if users don't pass the `offset` or `count` parameters, it should be
replaced with default value (0 and -1), but the previous behaviour is
trying to casting the parameters.
so, if one of them is `None`, the casting process will cause
`ValueError`...at least before python 3.8, i guess. currently it raise a
`TypeError` instead, update the API to catch this error and use default
value if get a `None`.